requirements.txt
numpy>=2.1.0
matplotlib>=3.10.0
opencv-python>=4.11.0
pillow>=12.0.0
mediapy>=1.2.5